+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 mongodb extension Change Log
 2.1.3 under development
 -----------------------
 
-- no changes in this release.
+- Bug #168: Fixed `yii\mongodb\Command::update()` uses `upsert` option by default (klimov-paul)
 
 
 2.1.2 October 31, 2016